Feat/added star rating for native manual 2 (#453)
Asana Link: https://app.asana.com/0/1203775898402819/1203512710344848

1. Add the google mediation adapter and the google ad manager adapter to
the gradle file, and then sync the project with the gradle files. There
might be an issue that pops up with the min sdk version, so you might
have to change the min sdk version of the project.
2. Add your "com.google.android.gms.ads.APPLICATION_ID" and your SDK Key
to the AndroidManifest.XML file.
3. Add your native ad unit identifier to the ManualNativeAdActivity and
the ManualNativeLateBindingAdActivity. Note that you need to do this in
order to get the full contents of the ad.
4. When loading and showing the test ads run Charles. I then went to
tools -> Map Local and used the attached file. Attached is also a
screenshot on how to configure the settings for the Map Local option
<img width="896" alt="image"
src="https://user-images.githubusercontent.com/123420316/217082129-569ae253-2a46-4b7a-8015-4ae8d9082774.png">

ScreenShots:


![Screenshot_20230206-122117](https://user-images.githubusercontent.com/123420316/217082728-05d680b2-4821-4137-bed5-7e902745a88d.png)


![Screenshot_20230206-122200](https://user-images.githubusercontent.com/123420316/217082736-6d511270-d9b3-4723-b251-6007ea470973.png)