ref/remove_outdated_comment (#816)
Is this an omission of correction during the following PR?
https://github.com/AppLovin/AppLovin-MAX-SDK-Android/pull/649

<!-- Korbit AI PR Description Start -->
## Description by Korbit AI

### What change is being made?
Remove outdated and incorrect statements related to testing an AppLovin
SDK key from the `GlobalApplication.java` file.

### Why are these changes being made?
The removed comments were outdated and potentially misleading regarding
the setup of the AppLovin SDK as the current process may differ, and
maintaining accurate and relevant documentation within the code is
essential for clarity and proper setup instructions.

> Is this description stale? Ask me to generate a new description by
commenting `/korbit-generate-pr-description`
<!-- Korbit AI PR Description End -->